+++ jdaviz/configs/default/plugins/plot_options/plot_options.py
@@ -187,6 +187,8 @@ class PlotOptions(PluginTemplateMixin):
     stretch_preset_value = Any().tag(sync=True)  # glue will pass either a float or string
     stretch_preset_sync = Dict().tag(sync=True)
 
+    stretch_vstep = Float(0.1).tag(sync=True)  # dynamic based on full range from image
+
     stretch_vmin_value = Float().tag(sync=True)
     stretch_vmin_sync = Dict().tag(sync=True)
 
@@ -621,6 +623,9 @@ class PlotOptions(PluginTemplateMixin):
             if len(sub_data) > 0:
                 hist_lims = interval.get_limits(sub_data)
                 hist_mark.min, hist_mark.max = hist_lims
+                stretch_vstep = (hist_lims[1] - hist_lims[0]) / 100.
+                # round to 1-2 significant digits to avoid random step sizes
+                self.stretch_vstep = np.round(stretch_vstep, decimals=-int(np.log10(stretch_vstep))+1)  # noqa
             hist_mark.bins = self.stretch_hist_nbins
             # in case only the sample has changed but its length has not,
             # we'll force the traitlet to trigger a change